+++ pbxplore/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .formats.loader import chains_from_files, chains_from_trajectory
