+++ pbxplore/analysis/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .clustering import hclust
+from .neq import compute_neq, write_neq
