+++ pbxplore/formats/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .PDB import PDB_EXTENSIONS, PDBx_EXTENSIONS
